feat: move ErrorAlerts to be more accessible to all editors
