+++ src/editors/containers/ErrorAlerts/UploadErrorAlert.jsx
@@ -3,10 +3,10 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from '@edx/frontend-platform/i18n';
 
-import messages from '../SelectImageModal/messages';
+import messages from '../TextEditor/components/SelectImageModal/messages';
 import ErrorAlert from './ErrorAlert';
-import { selectors } from '../../../../data/redux';
-import { RequestKeys } from '../../../../data/constants/requests';
+import { selectors } from '../../data/redux';
+import { RequestKeys } from '../../data/constants/requests';
 
 export const UploadErrorAlert = ({
   // redux
