+++ src/editors/containers/ErrorAlerts/UploadErrorAlert.test.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import { UploadErrorAlert, mapStateToProps } from './UploadErrorAlert';
-import { selectors } from '../../../../data/redux';
-import { RequestKeys } from '../../../../data/constants/requests';
+import { selectors } from '../../data/redux';
+import { RequestKeys } from '../../data/constants/requests';
 
-jest.mock('../../../../data/redux', () => ({
+jest.mock('../../data/redux', () => ({
   selectors: {
     requests: {
       isFailed: jest.fn((state, params) => ({ isFailed: { state, params } })),
